# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZARITH Executive AI Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  failed: "failed",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string | null;
  status: TaskStatus;
  priority: TaskPriority;
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateTaskBody {
  title: string;
  description?: string | null;
  priority?: CreateTaskBodyPriority;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  failed: "failed",
} as const;

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UpdateTaskBody {
  title?: string;
  description?: string | null;
  status?: UpdateTaskBodyStatus;
  priority?: UpdateTaskBodyPriority;
}

export interface SendMessageBody {
  content: string;
  sessionId?: string | null;
}

export type AgentMessageRole =
  (typeof AgentMessageRole)[keyof typeof AgentMessageRole];

export const AgentMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AgentMessage {
  id: number;
  role: AgentMessageRole;
  content: string;
  createdAt: string;
  isThinking: boolean;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export type DashboardSummaryAgentStatus =
  (typeof DashboardSummaryAgentStatus)[keyof typeof DashboardSummaryAgentStatus];

export const DashboardSummaryAgentStatus = {
  idle: "idle",
  thinking: "thinking",
  executing: "executing",
} as const;

export interface DashboardSummary {
  totalTasks: number;
  completedTasks: number;
  pendingTasks: number;
  inProgressTasks: number;
  totalMessages: number;
  agentStatus: DashboardSummaryAgentStatus;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  task_created: "task_created",
  task_completed: "task_completed",
  task_failed: "task_failed",
  message_sent: "message_sent",
  agent_response: "agent_response",
} as const;

export type ActivityItemMetadata = { [key: string]: unknown } | null;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  metadata?: ActivityItemMetadata;
  createdAt: string;
}

export type ListTasksParams = {
  status?: ListTasksStatus;
  limit?: number;
};

export type ListTasksStatus =
  (typeof ListTasksStatus)[keyof typeof ListTasksStatus];

export const ListTasksStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  failed: "failed",
} as const;

export type GetChatHistoryParams = {
  limit?: number;
};

export type GetRecentActivityParams = {
  limit?: number;
};