# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZARITH Executive AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all tasks
 */
export const listTasksQueryLimitDefault = 50;

export const ListTasksQueryParams = zod.object({
  status: zod
    .enum(["pending", "in_progress", "completed", "failed"])
    .optional(),
  limit: zod.coerce.number().default(listTasksQueryLimitDefault),
});

export const ListTasksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["pending", "in_progress", "completed", "failed"]),
  priority: zod.enum(["low", "medium", "high", "critical"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListTasksResponse = zod.array(ListTasksResponseItem);

/**
 * @summary Create a new task
 */
export const createTaskBodyPriorityDefault = `medium`;

export const CreateTaskBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
  priority: zod
    .enum(["low", "medium", "high", "critical"])
    .default(createTaskBodyPriorityDefault),
});

/**
 * @summary Get task by ID
 */
export const GetTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["pending", "in_progress", "completed", "failed"]),
  priority: zod.enum(["low", "medium", "high", "critical"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a task
 */
export const UpdateTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTaskBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().nullish(),
  status: zod
    .enum(["pending", "in_progress", "completed", "failed"])
    .optional(),
  priority: zod.enum(["low", "medium", "high", "critical"]).optional(),
});

export const UpdateTaskResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["pending", "in_progress", "completed", "failed"]),
  priority: zod.enum(["low", "medium", "high", "critical"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a task
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Send a message to the ZARITH agent
 */
export const SendMessageBody = zod.object({
  content: zod.string(),
  sessionId: zod.string().nullish(),
});

export const SendMessageResponse = zod.object({
  id: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  createdAt: zod.coerce.date(),
  isThinking: zod.boolean(),
});

/**
 * @summary Get chat message history
 */
export const getChatHistoryQueryLimitDefault = 50;

export const GetChatHistoryQueryParams = zod.object({
  limit: zod.coerce.number().default(getChatHistoryQueryLimitDefault),
});

export const GetChatHistoryResponseItem = zod.object({
  id: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const GetChatHistoryResponse = zod.array(GetChatHistoryResponseItem);

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardSummaryResponse = zod.object({
  totalTasks: zod.number(),
  completedTasks: zod.number(),
  pendingTasks: zod.number(),
  inProgressTasks: zod.number(),
  totalMessages: zod.number(),
  agentStatus: zod.enum(["idle", "thinking", "executing"]),
});

/**
 * @summary Get recent activity feed
 */
export const getRecentActivityQueryLimitDefault = 10;

export const GetRecentActivityQueryParams = zod.object({
  limit: zod.coerce.number().default(getRecentActivityQueryLimitDefault),
});

export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum([
    "task_created",
    "task_completed",
    "task_failed",
    "message_sent",
    "agent_response",
  ]),
  description: zod.string(),
  metadata: zod.object({}).passthrough().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);