# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ZARITH Executive AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryAgentStatus } from "./dashboardSummaryAgentStatus";

export interface DashboardSummary {
  totalTasks: number;
  completedTasks: number;
  pendingTasks: number;
  inProgressTasks: number;
  totalMessages: number;
  agentStatus: DashboardSummaryAgentStatus;
}